pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "drift-autopsy"
version = "0.1.0"
description = "Modular data drift detection, diagnosis, and remediation framework"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Nathan", email = "nathan@example.com"}
]
keywords = ["machine-learning", "drift-detection", "monitoring", "mlops"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "numpy>=1.21.0",
    "pandas>=1.3.0",
    "scipy>=1.7.0",
    "scikit-learn>=1.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "structlog>=23.1.0",
]

[project.optional-dependencies]
analysis = [
    "shap>=0.41.0",
    "matplotlib>=3.5.0",
    "seaborn>=0.12.0",
]
demo = [
    "folktables>=0.0.11",
    "jupyter>=1.0.0",
    "notebook>=6.5.0",
]
dashboard = [
    "streamlit>=1.28.0",
    "plotly>=5.17.0",
]
all = ["drift-autopsy[analysis,demo,dashboard]"]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
    "black>=23.0.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --tb=short"

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.black]
line-length = 100
target-version = ['py38']

[tool.ruff]
line-length = 100
target-version = "py38"
